Check that sum of components is not greater than total votes.
Depending on what the counties report, it is often the case when the sum
of the components is less than the total votes.  As a result, we will
simply test for the case where the sum of the components is greater than
the total votes.

This also adds some additional components ('absentee', 'polling') to
consider.