Ignore polling component.
This appears to be equivalent to 'election_day', which is the preferred
header name.  While only a few repositories have files with a 'polling'
header, there are a few that have both 'election_day' and 'polling',
which resulted in a false positive.  To simplify the tests, we will only
consider 'election_day'.